+++ main.ts
@@ -1,10 +1,25 @@
 import { Plugin } from 'obsidian';
 import { TikzModal } from './src/modal';
+import { DEFAULT_PLUGIN_DATA, PluginData, UserTemplate } from './src/templates';
 
-export default class SimpleTikzPlugin extends Plugin {
+export default class EasyTikzPlugin extends Plugin {
+    data: PluginData = { ...DEFAULT_PLUGIN_DATA };
+
     async onload() {
-        this.addRibbonIcon('square-function', 'TikZ Graph Helper', () => {
-            new TikzModal(this.app).open();
+        const stored = (await this.loadData()) as Partial<PluginData> | null;
+        this.data = {
+            ...DEFAULT_PLUGIN_DATA,
+            ...(stored ?? {}),
+            userTemplates: Array.isArray(stored?.userTemplates) ? stored!.userTemplates : [],
+        };
+
+        this.addRibbonIcon('square-function', 'Easy TikZ', () => {
+            new TikzModal(this.app, this).open();
         });
+    }
+
+    async saveUserTemplates(templates: UserTemplate[]) {
+        this.data.userTemplates = templates;
+        await this.saveData(this.data);
     }
 }